LayoutTests/editing/pasteboard/simplfiying-markup-should-not-strip-content-expected.txt

This tests copying and pasting doesn't strip content.
To manually test, copy the content in the first box and paste it into the second box.

Original content:
| "\n"
| <font>
|   face="Verdana"
|   "hello "
| <br>
| "\n"
| <font>
|   face="Verdana"
|   <div>
|     style="-webkit-line-break: after-white-space;"
|     <div>
|       <font>
|         face="Verdana"
|         "world"
|     "\n"
|     <div>
|       style="-webkit-line-break: after-white-space; "
|       <font>
|         face="Verdana"
|         "WebKit"
|   "\n"

Pasted content:
| <font>
|   face="Verdana"
|   "hello "
| <br>
|   style="font-family: Helvetica;"
| <font>
|   face="Verdana"
|   "worldWebKit<#selection-caret>"
